% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaseAccessibility.R
\docType{data}
\name{CreateBaseAccessibilitySpecifications}
\alias{CreateBaseAccessibilitySpecifications}
\title{Specifications list for CreateBaseAccessibility module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CreateBaseAccessibility.R script.
}
\usage{
CreateBaseAccessibilitySpecifications
}
\description{
A list containing specifications for the CreateBaseAccessibility module.
}
\keyword{datasets}
