% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFutureAccessibility.R
\name{CreateFutureAccessibility}
\alias{CreateFutureAccessibility}
\title{Calculate freeway lane-miles per capita by Marea and
calculate bus and rail revenue miles by Marea.}
\usage{
CreateFutureAccessibility(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CreateFutureAccessibility} calculate freeway and arterial lane-miles per capita, and
bus and rail rev-miles.
}
\details{
This function calculates freeway and arterial lane-miles per capita, and
bus and rail rev-miles for each Marea.
}
