% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTransitService.R
\name{estimateTransitModel}
\alias{estimateTransitModel}
\title{Estimate public transit model parameters.}
\usage{
estimateTransitModel()
}
\value{
A list containing the following elements:
BusEquivalents_df: factors to convert revenue miles by mode into bus
equivalents,
UZABusEqRevMile_df: data on bus equivalent revenue miles by urbanized area,
VehMiFactors_df: factors to convert revenue miles by mode into vehicle miles
by mode.
}
\description{
\code{estimateTransitModel} estimates transit model parameters.
}
\details{
This function estimates transit model parameters from 2015 National Transit
Database information on transit agencies and service levels. The function
calculates factors for converting annual revenue miles by transit mode to
total bus-equivalent revenue miles. It also calculates factors to convert
revenue miles by mode into vehicle miles by mode.
}
