% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTransitService.R
\docType{data}
\name{UZABusEqRevMile_df}
\alias{UZABusEqRevMile_df}
\title{Urbanized area bus equivalent revenue mile data for 2015}
\format{A data frame with 439 rows and 5 variables containing urbanized area
data on bus equivalent revenue miles

\describe{
  \item{UzaCode}{integer code corresponding to 5-digit code used in the NTD}
  \item{UzaName}{urbanized area name}
  \item{BusEqRevMi}{annual bus equivalent revenue miles in the urbanized area}
  \item{UzaPop}{urbanized area population}
  \item{BusEqRevMiPC}{annual bus equivalent revenue miles per capita in the urbanized area}
}}
\source{
AssignTransitService.R script.
}
\usage{
UZABusEqRevMile_df
}
\description{
Urbanized area data from the 2015 National Transit Database (NTD) related to
the calculation of bus equivalent revenue miles and per capita values.
}
\keyword{datasets}
